backend/pyproject.toml
[tool.poetry]
name = "code-of-the-world"
version = "0.1.0"
description = ""
authors = ["kirill <shkolenko02-03-2004@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.0"}
alembic = "^1.11.3"
loguru = "^0.7.0"
ruff = "^0.3.7"
pre-commit = "^3.7.0"
llama-cpp-python = "^0.2.76"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
